Refactor : FinalFrameState 제거 로직 이관
